PAC-Bench/scenario_simulator/agent_service/requirements.txt
langchain
langchain-mcp-adapters
python-dotenv
langchain-anthropic
langchain-openai
fastapi
langgraph